config: Move config.py to config/keys.py & update imports.
